Firebase: Login e registro de adminstradores concluídos

+++ src/app/firebaseconfig.ts
@@ -7,19 +7,22 @@ import {
   updateDoc,
   doc,
   getDocs,
-  DocumentReference
+  DocumentReference,
+  onSnapshot,
 } from "firebase/firestore";
+import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword, signOut } from "firebase/auth";
 
 const firebaseConfig = {
-  apiKey: "AIzaSyCbWio8rerVcId4k3WGDhrKuahOj2t2v9I",
-  authDomain: "app-fitfusion.firebaseapp.com",
-  projectId: "app-fitfusion",
-  storageBucket: "app-fitfusion.appspot.com",
-  messagingSenderId: "61815565821",
-  appId: "1:61815565821:web:52334725ccd5c6896bbd94",
+  apiKey: process.env.NEXT_PUBLIC_FIREBASE_API_KEY,
+  authDomain: process.env.NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN,
+  projectId: process.env.NEXT_PUBLIC_FIREBASE_PROJECT_ID,
+  storageBucket: process.env.NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET,
+  messagingSenderId: process.env.NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID,
+  appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
 };
 
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
+const auth = getAuth(app);
 
-export { db, collection, addDoc, doc, deleteDoc, updateDoc, getDocs, DocumentReference };
+export { auth, db, collection, addDoc, doc, deleteDoc, updateDoc, getDocs, DocumentReference, onSnapshot };
